Ensure remote command ids can be 64-bit integers
Bug: b/280045887
Cq-Include-Trybots: luci.chrome.try:linux-chromeos-chrome
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4489067
Quick-Run: Jeroen Dhollander <jeroendh@google.com>
Auto-Submit: Jeroen Dhollander <jeroendh@google.com>
Commit-Queue: Pavol Marko <pmarko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1146492}

+++ components/policy/test_support/remote_commands_result_waiter.h
@@ -6,6 +6,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef COMPONENTS_POLICY_TEST_SUPPORT_REMOTE_COMMANDS_RESULT_WAITER_H_
 #define COMPONENTS_POLICY_TEST_SUPPORT_REMOTE_COMMANDS_RESULT_WAITER_H_
 
+#include <cstdint>
+
 #include "base/memory/raw_ptr.h"
 #include "base/run_loop.h"
 #include "components/policy/proto/device_management_backend.pb.h"
@@ -17,7 +19,7 @@ namespace policy {
 class RemoteCommandsResultWaiter : public RemoteCommandsState::Observer {
  public:
   RemoteCommandsResultWaiter(RemoteCommandsState* remote_commands_state,
-                             int command_id);
+                             int64_t command_id);
 
   ~RemoteCommandsResultWaiter() override;
 
@@ -26,10 +28,10 @@ class RemoteCommandsResultWaiter : public RemoteCommandsState::Observer {
   enterprise_management::RemoteCommandResult WaitAndGetResult();
 
  private:
-  void OnRemoteCommandResultAvailable(int command_id) override;
+  void OnRemoteCommandResultAvailable(int64_t command_id) override;
 
   const raw_ptr<RemoteCommandsState> remote_commands_state_;
-  const int command_id_;
+  const int64_t command_id_;
   base::RunLoop run_loop_;
 };
